% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argentina_STRs.R
\docType{data}
\name{Argentina_STRs}
\alias{Argentina_STRs}
\title{STRs allelic frequencies from Argentina.}
\format{
A data frame with allele frequencies
}
\usage{
Argentina_STRs
}
\description{
STRs allelic frequencies from Argentina.
}
\keyword{datasets}
