% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnet.R
\docType{data}
\name{bnet}
\alias{bnet}
\title{Initialized bayesian network.}
\format{
Initialized bayesian network, based on toyped.
}
\usage{
bnet
}
\description{
Initialized bayesian network.
}
\keyword{datasets}
