% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBN.R
\name{buildBN}
\alias{buildBN}
\title{buildBN: a function for building the bayesian network.}
\usage{
buildBN(pbn, QP)
}
\arguments{
\item{pbn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{QP}{Query Persons Ids}
}
\value{
A bayesian network based on pedigree evidence and QP definition.
}
\description{
buildBN: a function for building the bayesian network.
}
\examples{
pbn  <- initBN(toyped)
bnet <- buildBN(pbn,QP=3)
}
