% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPedformat.R
\name{convertPedformat}
\alias{convertPedformat}
\title{convertPedformat: a function for converting a pedtools ped onject to a famlink ped object.}
\usage{
convertPedformat(x, verbose = FALSE)
}
\arguments{
\item{x}{A pedtools ped object.}

\item{verbose}{Function output.}
}
\value{
A dataframe with LRs.
}
\description{
convertPedformat: a function for converting a pedtools ped onject to a famlink ped object.
}
