% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidencePrunning.R
\name{evidencePrunning}
\alias{evidencePrunning}
\title{evidencePrunning: a fuction for pruning instantiated variables.}
\usage{
evidencePrunning(bn)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}
}
\value{
A preprocessed bayesian network.
}
\description{
evidencePrunning: a fuction for pruning instantiated variables.
}
