% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHeteroFounders.R
\name{factorHeteroFounders}
\alias{factorHeteroFounders}
\title{factorHeteroFounders: a function for multiplying probabilities in case of heterocigote founders.}
\usage{
factorHeteroFounders(rresQ, bn)
}
\arguments{
\item{rresQ}{List of CPTs.}

\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}
}
\value{
A dataframe with genotype probabilities.
}
\description{
factorHeteroFounders: a function for multiplying probabilities in case of heterocigote founders.
}
