% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamiliasPedigree.R
\name{FamiliasPedigree}
\alias{FamiliasPedigree}
\title{FamiliasPedigree: a function for constructing Familias pedigree format.}
\usage{
FamiliasPedigree(id, dadid, momid, sex)
}
\arguments{
\item{id}{individual id}

\item{dadid}{father id}

\item{momid}{mother id}

\item{sex}{biological sex}
}
\value{
A dataframe with probabilities.
}
\description{
FamiliasPedigree: a function for constructing Familias pedigree format.
}
