% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnet.R
\docType{package}
\name{fbnet}
\alias{fbnet}
\title{fbnet: Forensic Bayesian Networks}
\description{
'fbnet' is an open source software package written in R statistical languaje.
It relies on a Bayesian Networks framework \doi{10.1017/CBO9780511811357}. It is particularly well suited
to efficiently perform large-size queries against databases of missing individuals.
It could interact with the main functionallities of other packages for pedigree analysis.
In particular, 'fbnet' imports the 'Familias' software
\doi{10.1016/S0379-0738(00)00147-X}. In addition 'pedtools', a software for creating
and manipulating pedigrees and markers, is supported. 'fbnet' allows computing LRs
and obtaining genotype probability distributions for query individual, based on
the pedigree data.
'fbnet' implements the complete GENis functionality, a recently published open-source
multi-tier information system developed to run forensic DNA databases
to perform kinship analysis based on DNA profiles \doi{10.1016/j.fsir.2020.100132}.
}
