% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditional.R
\name{getConditional}
\alias{getConditional}
\title{getConditional: a function for obtaining the coditional probability tables based on a given evidence.}
\usage{
getConditional(lf)
}
\arguments{
\item{lf}{A list of joint probabilities.}
}
\value{
A list of conditioned probabilities.
}
\description{
getConditional: a function for obtaining the coditional probability tables based on a given evidence.
}
