% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenotypeTables.R
\name{getGenotypeTables}
\alias{getGenotypeTables}
\title{getGenotypeTables: a function for obtaining genotypetables after variable elimination and using available genetic evidence.}
\usage{
getGenotypeTables(bn, resQ, geno = NULL, lqp = NULL)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{resQ}{List of CPTs.}

\item{geno}{data.frame with genotypes.}

\item{lqp}{list of individuals genotypes.}
}
\value{
A dataframe with genotype probabilities.
}
\description{
getGenotypeTables: a function for obtaining genotypetables after variable elimination and using available genetic evidence.
}
