% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocusCPT.R
\name{getLocusCPT}
\alias{getLocusCPT}
\title{getLocusCPT: a function for obtaining the coditional probability table from a specific locus.}
\usage{
getLocusCPT(bn, locus, lumpingParameter = NULL, renorm = "row-wise")
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{locus}{Specified locus.}

\item{lumpingParameter}{Used for stepwise mutational model.}

\item{renorm}{If "row-wise" is selected, zero probability is assigned for transitions out of range.}
}
\value{
A bayesian network based on pedigree evidence and QP definition.
}
\description{
getLocusCPT: a function for obtaining the coditional probability table from a specific locus.
}
\examples{
pbn  <- initBN(toyped)
bnet <- buildBN(pbn,QP=3)
bn1  <- buildCPTs(bnet)
locCPT <- getLocusCPT(bn1,"M1")
}
