% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAP.R
\name{getMAP}
\alias{getMAP}
\title{factorHeteroFounders: a function for multiplying probabilities in case of heterocigote founders.}
\usage{
getMAP(resQ, topn = 3)
}
\arguments{
\item{resQ}{List of CPTs.}

\item{topn}{Format parameter.}
}
\value{
A MAP from the probability table.
}
\description{
factorHeteroFounders: a function for multiplying probabilities in case of heterocigote founders.
}
