% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQSetRMP.R
\name{getQSetRMP}
\alias{getQSetRMP}
\title{getGenotypeTables: a function for obtaining genotypetables after variable elimination and using available genetic evidence.}
\usage{
getQSetRMP(bn, lqp)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{lqp}{list of individuals genotypes.}
}
\value{
A dataframe with genotype probabilities.
}
\description{
getGenotypeTables: a function for obtaining genotypetables after variable elimination and using available genetic evidence.
}
