% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValuesOut.R
\name{getValuesOut}
\alias{getValuesOut}
\title{getValuesOut: a function for getting out variables with zero probability in the bayesian network}
\usage{
getValuesOut(cpt, condVar = c())
}
\arguments{
\item{cpt}{conditional probability table from the bayesian network}

\item{condVar}{variables from the conditioning table}
}
\value{
A processed conditional probability table
}
\description{
getValuesOut: a function for getting out variables with zero probability in the bayesian network
}
