% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeEvidence.R
\name{imposeEvidence}
\alias{imposeEvidence}
\title{imposeEvidence: a fuction for imposing evidence in the bayesian network.}
\usage{
imposeEvidence(bn)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}
}
\value{
A preprocessed bayesian network.
}
\description{
imposeEvidence: a fuction for imposing evidence in the bayesian network.
}
