% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBN.fromPed.R
\name{initBN.fromPed}
\alias{initBN.fromPed}
\title{initBN.fromPed: a function to initialize the bayesian network.}
\usage{
initBN.fromPed(ped, bplotped)
}
\arguments{
\item{ped}{A ped object in Familias format.}

\item{bplotped}{An alternative ped object to be compared.}
}
\value{
A bayesian network.
}
\description{
initBN.fromPed: a function to initialize the bayesian network.
}
