% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBN.fromVars.R
\name{initBN.fromVars}
\alias{initBN.fromVars}
\title{initBN.fromVars: a function to initialize the bayesian network.}
\usage{
initBN.fromVars(bplotped)
}
\arguments{
\item{bplotped}{An alternative ped object to be compared.}
}
\value{
A bayesian network.
}
\description{
initBN.fromVars: a function to initialize the bayesian network.
}
