% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBN.R
\name{initBN}
\alias{initBN}
\title{initBN: a function to initialize the bayesian network.}
\usage{
initBN(ped = NULL, bplotped = FALSE)
}
\arguments{
\item{ped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}

\item{bplotped}{An alternative ped object to be compared.}
}
\value{
A bayesian network.
}
\description{
initBN: a function to initialize the bayesian network.
}
\examples{
ped <- list(
  id = c(1, 2, 3),
  father = c(0, 0, 1),
  mother = c(0, 0, 2),
  sex = c(1, 2, 1),  # 1 = male, 2 = female
  genotype = c("AA", "AB", "BB")
)
bplotped <- list(
  id = c(4, 5),
  father = c(0, 4),
  mother = c(0, 5),
  sex = c(1, 2),
  genotype = c("AB", "AA")
)
}
