% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.linkdat.R
\name{is.linkdat}
\alias{is.linkdat}
\alias{is.singletonfbfb}
\alias{is.linkdat.list}
\title{Is an object a linkdat object?}
\usage{
is.linkdat(x)

is.singletonfbfb(x)

is.linkdat.list(x)
}
\arguments{
\item{x}{Any \code{R} object.}
}
\value{
For \code{is.linkdat}: TRUE if \code{x} is a linkdat (or singletonfb)
object, and FALSE otherwise.\cr For \code{is.singletonfbfb}: TRUE if \code{x}
is a singletonfb object, and FALSE otherwise.\cr For \code{is.linkdat.list}:
TRUE if \code{x} is a list of linkdat/singletonfb objects.
}
\description{
Functions for checking whether an object is a \code{\link{linkdat}} object, a
\code{\link{singletonfb}} or a list of such.
}
\details{
Note that the \code{singletonfb} class inherits from \code{linkdat}, so if
\code{x} is a singletonfb, \code{is.linkdat(x)} returns TRUE.
}
\seealso{
\code{\link{linkdat}}
}
