% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendelianCheckfb.R
\name{mendelianCheckfb}
\alias{mendelianCheckfb}
\title{Check for Mendelian errors}
\usage{
mendelianCheckfb(x, remove = FALSE, verbose = !remove)
}
\arguments{
\item{x}{a linkdat object}

\item{remove}{a logical. If FALSE, the function returns the indices of
markers found to incorrect.  If TRUE, a new linkdat object is
returned, where the incorrect markers have been deleted.}

\item{verbose}{a logical. If TRUE, details of the markers failing the tests
are shown.}
}
\value{
A numeric containing the indices of the markers that did not pass the
tests, or (if \code{remove=TRUE}) a new \code{linkdat} object where the
failing markers are removed.
}
\description{
Check marker data for Mendelian inconsistencies
}
