% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minOrdering.R
\name{minOrdering}
\alias{minOrdering}
\title{minOrdering: a function for getting an ordering of bayesian network variables not in Q using min fill criteria on interaction graphs.}
\usage{
minOrdering(bn, vars = NULL, method = c("min_degree", "min_fill")[1])
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{vars}{Subset of tables where the order is calculated}

\item{method}{Elimination method, min_degree or min_fill}
}
\value{
A bayesian network after ordering process.
}
\description{
minOrdering: a function for getting an ordering of bayesian network variables not in Q using min fill criteria on interaction graphs.
}
