% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\docType{data}
\name{pbn}
\alias{pbn}
\title{Prepared pedigree for bayesian network trimming.}
\format{
A data.matrix.
}
\usage{
pbn
}
\description{
Prepared pedigree for bayesian network trimming.
}
\keyword{datasets}
