% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePed.R
\name{preparePed}
\alias{preparePed}
\title{preparePed: a function for simulating genetic data from untyped individuals conditioned on known genotypes.}
\usage{
preparePed(ped, available, lLociFreq, rseed = NULL)
}
\arguments{
\item{ped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}

\item{available}{Genotyped individuals IDs.}

\item{lLociFreq}{Allele frequencies.}

\item{rseed}{Seed used for simulations.}
}
\value{
A ped object.
}
\description{
preparePed: a function for simulating genetic data from untyped individuals conditioned on known genotypes.
}
