% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodFactor.R
\name{prodFactor}
\alias{prodFactor}
\title{prodFactor: a function for performing product between probability tables.}
\usage{
prodFactor(laux)
}
\arguments{
\item{laux}{probability distribution aux}
}
\value{
A dataframe with probabilities.
}
\description{
prodFactor: a function for performing product between probability tables.
}
