% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneNodes.R
\name{pruneNodes}
\alias{pruneNodes}
\title{pruneNodes: a fuction for clasical pruning in bayesian networks.}
\usage{
pruneNodes(bn)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}
}
\value{
A preprocessed bayesian network.
}
\description{
pruneNodes: a fuction for clasical pruning in bayesian networks.
}
