% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEvidenceFromPed.R
\name{removeEvidenceFromPed}
\alias{removeEvidenceFromPed}
\title{removeEvidenceFromPed: a function for removing evidence from specific individuals in a ped object.}
\usage{
removeEvidenceFromPed(pped, idNotEv)
}
\arguments{
\item{pped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}

\item{idNotEv}{A set of individuals whom evidence should be removed.}
}
\value{
A ped object.
}
\description{
removeEvidenceFromPed: a function for removing evidence from specific individuals in a ped object.
}
