% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLR.R
\name{reportLR}
\alias{reportLR}
\title{reportLR: a function for calculating the LRs of specified genotypes in a pedigree.}
\usage{
reportLR(bn, resQ, geno = NULL)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{resQ}{List of CPTs.}

\item{geno}{data.frame with genotypes.}
}
\value{
A dataframe with LRs.
}
\description{
reportLR: a function for calculating the LRs of specified genotypes in a pedigree.
}
