% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPQ.R
\name{reportPQ}
\alias{reportPQ}
\title{reportPQ: a function for calculating the probability of specified genotypes in a pedigree.}
\usage{
reportPQ(bn, resQ, geno = NULL)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{resQ}{List of CPTs.}

\item{geno}{data.frame with genotypes.}
}
\value{
A dataframe with genotype probabilities.
}
\description{
reportPQ: a function for calculating the probability of specified genotypes in a pedigree.
}
