% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseSplit.R
\name{reverseSplit}
\alias{reverseSplit}
\title{reverseSpit: a function for formatting.}
\usage{
reverseSplit(inList)
}
\arguments{
\item{inList}{input for formatting.}
}
\value{
A bayesian network.
}
\description{
reverseSpit: a function for formatting.
}
