% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFamilias.R
\name{runFamilias}
\alias{runFamilias}
\title{runFamilias: a function for running Familias likelihood ratio computations.}
\usage{
runFamilias(ped = NULL)
}
\arguments{
\item{ped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}
}
\value{
Computation results.
}
\description{
runFamilias: a function for running Familias likelihood ratio computations.
}
