% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFamilias_withMutations.R
\name{runFamilias_withMutations}
\alias{runFamilias_withMutations}
\title{runFamilias: a function for running Familias likelihood ratio computations.}
\usage{
runFamilias_withMutations(
  ped = NULL,
  mutModel = c("Equal", "Stepwise")[2],
  R = 0.005,
  R2 = 5e-07,
  r = 0.5,
  maleMutModel = NULL,
  maleR = NULL,
  maleR2 = NULL,
  male.r = NULL,
  femaleMutModel = NULL,
  femaleR = NULL,
  femaleR2 = NULL,
  female.r = NULL
)
}
\arguments{
\item{ped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}

\item{mutModel}{mutational models supported from Familias, options are: "equal", "Stepwise".}

\item{R}{mutational rate}

\item{R2}{mutational rate for microvariants}

\item{r}{range (used in stepwise model)}

\item{maleMutModel}{male model}

\item{maleR}{male R}

\item{maleR2}{male R2}

\item{male.r}{male range}

\item{femaleMutModel}{female model}

\item{femaleR}{female R}

\item{femaleR2}{femaleR2}

\item{female.r}{female range}
}
\value{
Computation results.
}
\description{
runFamilias: a function for running Familias likelihood ratio computations.
}
