% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedModify.R
\name{setAvailable}
\alias{setAvailable}
\alias{swapAvailable}
\title{Functions for modifying availability vectors}
\usage{
setAvailable(x, available)

swapAvailable(x, ids)
}
\arguments{
\item{x}{a \code{linkdat} object}

\item{available}{a numeric containing the IDs of available individuals.}

\item{ids}{individuals}
}
\value{
The modified \code{linkdat} object.
}
\description{
Functions to set and modify the availability vector of a 'linkdat' object.
This vector is used in 'linkage.power' and 'linkageSim', indicating for whom
genotypes should be simulated.
}
\seealso{
\code{plot.linkdat}, \code{linkage.power},
\code{linkageSim}
}
