% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOrdering.R
\name{setOrdering}
\alias{setOrdering}
\title{setOrdering: a function for selecting the ordering method in the elimination process.}
\usage{
setOrdering(bn, ordMethod, vars = NULL, orderElim = NULL)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{ordMethod}{Ordering method.}

\item{vars}{Vars}

\item{orderElim}{Order elimination criteria.}
}
\value{
A bayesian network after ordering process.
}
\description{
setOrdering: a function for selecting the ordering method in the elimination process.
}
