% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateRemoval.R
\name{stateRemoval}
\alias{stateRemoval}
\title{stateRemoval: a function for processing the bayesian network.}
\usage{
stateRemoval(bn)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}
}
\value{
A preprocessed bayesian network.
}
\description{
stateRemoval: a function for processing the bayesian network.
}
