% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateRemoval2.R
\name{stateRemoval2}
\alias{stateRemoval2}
\title{stateRemoval2: a function for processing the bayesian network. It implements another approach from the described in stateRemoval function.}
\usage{
stateRemoval2(bn, verbose = FALSE)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}

\item{verbose}{Computation output.}
}
\value{
A preprocessed bayesian network.
}
\description{
stateRemoval2: a function for processing the bayesian network. It implements another approach from the described in stateRemoval function.
}
