% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateRemovalSubnucs.R
\name{stateRemovalSubnucs}
\alias{stateRemovalSubnucs}
\title{stateRemovalSubnucs: a fuctiong for variable state pruning.}
\usage{
stateRemovalSubnucs(bn, verbose = FALSE)
}
\arguments{
\item{bn}{A bayesian network (output of buildBN function).}

\item{verbose}{Computation output.}
}
\value{
A preprocessed bayesian network.
}
\description{
stateRemovalSubnucs: a fuctiong for variable state pruning.
}
