% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFactor.R
\name{sumFactor}
\alias{sumFactor}
\title{prodFactor: a function for performing sum between probability tables.}
\usage{
sumFactor(cpt, Z)
}
\arguments{
\item{cpt}{Conditional probability table}

\item{Z}{factor}
}
\value{
A dataframe with probabilities.
}
\description{
prodFactor: a function for performing sum between probability tables.
}
