% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toybase.R
\docType{data}
\name{toybase}
\alias{toybase}
\title{Toy allele frequency database.}
\format{
A data frame two markers allele frequencies
}
\usage{
toybase
}
\description{
Toy allele frequency database.
}
\keyword{datasets}
