% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyped.R
\docType{data}
\name{toyped}
\alias{toyped}
\title{STRs allelic frequencies from specified country.}
\format{
A toy pedigree. Nuclear family.
}
\usage{
toyped
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
