% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velim.bn.R
\name{velim.bn}
\alias{velim.bn}
\title{velim.bn: a function for variable elimination in a bayesian network.}
\usage{
velim.bn(
  bn,
  ordMethod = c("id", "min_degree", "min_fill", "fixed")[2],
  orderElim = NULL,
  verbose = FALSE
)
}
\arguments{
\item{bn}{A bayesian network for pedigree object with information of the genotyped members. The ped object must be in Familias format.}

\item{ordMethod}{Selected ordering method between id, min_degree, min_fill and fixed.}

\item{orderElim}{Elimination order.}

\item{verbose}{Computation output.}
}
\value{
Variable elimination result.
}
\description{
velim.bn: a function for variable elimination in a bayesian network.
}
\examples{
pbn  <- initBN(toyped)
bnet <- buildBN(pbn,QP=3)
bn1  <- buildCPTs(bnet)
resQ <- velim.bn(bn1,ordMethod="min_fill",verbose=FALSE)
}
