% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBN.R
\name{initBN}
\alias{initBN}
\title{initBN: a function to initialize the bayesian network.}
\usage{
initBN(ped = NULL, bplotped = FALSE)
}
\arguments{
\item{ped}{A ped object with information of the genotyped members. The ped object must be in Familias format.}

\item{bplotped}{An alternative ped object to be compared.}
}
\value{
A bayesian network.
}
\description{
initBN: a function to initialize the bayesian network.
}
\examples{
pbn  <- initBN(toyped)
}
